+++ CHANGELOG.md
@@ -4,8 +4,11 @@
 
 * M365DSCDRGUtil
   * Add `collectionId` export to `ConvertFrom-IntunePolicyAssignment`
+  * Add handling for Intune assignments in `Compare-M365DSCComplexObject`
+* M365DSCResourceGenerator
+  * Update CimInstance comparison template
 * MISC
-  * Add group display name export across Intune resources
+  * Add group display name export and update assignment comparison across Intune resources
 
 # 1.24.612.1